+++ src/main/java/Spring/MindStone/domain/member/MemberInterest.java
@@ -1,0 +1,56 @@
+package Spring.MindStone.domain.member;
+
+import Spring.MindStone.domain.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.DynamicInsert;
+import org.hibernate.annotations.DynamicUpdate;
+
+@Entity
+@Getter
+@Builder
+@DynamicUpdate
+@DynamicInsert
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class MemberInterest extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id; // ID (PK)
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private MemberInfo member; // 회원 ID (FK, Member와 관계 설정)
+
+    /*@Column(nullable = false)
+    private Integer actionID; // 행동 번호(취미, 특기, 스트레스 풀이)
+
+    @Column(nullable = false, columnDefinition = "TEXT")
+    private String action; // 행동 (텍스트)
+
+    @Column(nullable = false,columnDefinition = "INTEGER DEFAULT 0")
+    private Integer actionNum = 0; // 행동 횟수(회원이 얼마나 이 행동을 택했는지)*/
+
+
+    // 취미 선택 항목들
+    /*@ElementCollection -> 테이블 분기를 통한 유연한 관리 고려
+    @CollectionTable("")*/
+    @Column(name = "hobby_actions")
+    private String hobbyActions;   // 취미 행동 (운동, 독서 등)
+
+    @Column(name = "hobby_action_count")
+    private Integer hobbyActionCount;   // 취미 선택 횟수
+
+    @Column(name = "special_skill_actions")
+    private String specialSkillActions; // 특기 행동 (글쓰기, 요리 등)
+
+    @Column(name = "special_skill_action_count")
+    private Integer specialSkillActionCount;    // 특기 선택 횟수
+
+    @Column(name = "stress_actions")
+    private String stressActions;       // 스트레스 풀이 방법(행동)
+
+    @Column(name = "stress_actions_count")
+    private Integer stressActionsCount; // 스트레스 풀이 선택 횟수
+}
